+++ src/app/src/reducers/FeatureFlagsReducer.js
@@ -1,4 +1,5 @@
 import { createReducer } from 'redux-act';
+import update from 'immutability-helper';
 
 import {
     startFetchFeatureFlags,
@@ -7,11 +8,21 @@ import {
     clearFeatureFlags,
 } from '../actions/featureFlags';
 
-const initialState = Object.freeze({});
+const initialState = Object.freeze({
+    fetching: false,
+    flags: Object.freeze({}),
+});
 
 export default createReducer({
-    [startFetchFeatureFlags]: () => initialState,
-    [failFetchFeatureFlags]: () => initialState,
+    [startFetchFeatureFlags]: state => update(state, {
+        fetching: { $set: true },
+    }),
+    [failFetchFeatureFlags]: state => update(state, {
+        fetching: { $set: false },
+    }),
     [clearFeatureFlags]: () => initialState,
-    [completeFetchFeatureFlags]: (_, payload) => Object.freeze(payload),
+    [completeFetchFeatureFlags]: (state, data) => Object.freeze(update(state, {
+        fetching: { $set: false },
+        flags: { $set: data },
+    })),
 }, initialState);